Fix tar unpack problems in nix-init.sh
Tar extract have a problem creating symlinks in read-only directories. (It first
creates a regular file, changes the permission of the directory, and then
realizes that the file should have been a symlink, but at this point can no
longer remove it.) The --delay-directory-restore flag defers the permission
changes on the directories and fixes this issue.