New better way of identifying packages.
From now on, binary wrappers explicitly have to write out the nixpkgs
version used to install the package, the package name (attrpath), the
derivation hash and the out hash.  This makes it possible to
build (and verify) the package even if it gets deleted from the binary
cache on nixos.org.