Revert "Upgrade nix to 1.6.1"
This reverts commit 282224ae3351bcc32074456b42436b6b6da69cb7.

Unfortunately tar --keep-old-files is silly, the upcoming
replacment of --skip-old-files is only available from tar 1.27.
Tar 1.27 is very new, so we have to use it from Nix, it's just
not there on wheezy or any Ubuntu version.  It's not even there
in Nixpkgs (pull request sent).  Once we have it in Nixpkgs,
we have to fix ceh-upgrade.sh to use it from there and then
retry this upgrade.

:(